metodo flowerShop casteado

+++ src/main/java/org/example/FlowerShop.java
@@ -20,13 +20,13 @@ public class FlowerShop {
     }
     public void printQuantity (){
 
-        int  treeCount =  productList.stream().filter (object -> object instanceof Tree).count();
+        int  treeCount = (int) productList.stream().filter (object -> object instanceof Tree).count();
             System.out.println("There are  " + treeCount + " trees");
 
-        int  flowersCount =  productList.stream().filter (object -> object instanceof Flower).count();
+        int  flowersCount = (int) productList.stream().filter (object -> object instanceof Flower).count();
             System.out.println("There are  " + flowersCount + "flowers");
 
-        int  decorationCount =  productList.stream().filter (object -> object instanceof Decoration).count();
+        int  decorationCount = (int) productList.stream().filter (object -> object instanceof Decoration).count();
         System.out.println("There are  " + decorationCount + "decorations");
 
     }
